requirements.txt
# Core App
streamlit
pandas
plotly
openpyxl
xlsxwriter
pillow
fuzzywuzzy==0.18.0
python-Levenshtein

# Machine Learning / Fuzzy Matching
rapidfuzz

# Google Cloud / Drive Integration
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
# AI & Networking
requests
aiohttp
openai
google-generativeai
streamlit-autorefresh
